reverting changes due to userUtils
